requirements.txt
streamlit>=1.40.0
pandas>=2.0.3
numpy>=1.24.3
python-binance>=1.0.17
dateparser>=1.1.8
streamlit-authenticator==0.4.1
pyyaml>=6.0.1
bcrypt>=4.0.1
